Some rules are very simple
Some rules are very simple, as child(Y, X) :- parent(X, Y). which says that Y is the child of X in case X is the parent of Y. Other rules get complicated, like the rule for sibling(X, Y), where X and Y are siblings if the have the same father Z and the same mother W, and are not the same person. (Without the last, a person would be a sibling of themselves.